chore: move direct reference service
